requirements_supabase.txt
# Supabase Setup Requirements
# Install these packages for the Epicure Supabase integration

# Core Supabase client
supabase==2.4.2

# Environment variables
python-dotenv==1.0.0

# Database and vectors
psycopg2-binary==2.9.9
numpy==1.24.3

# HTTP requests (for setup script)
requests==2.31.0

# Optional: Direct PostgreSQL access
sqlalchemy==2.0.23
asyncpg==0.29.0

# AI/ML integrations
openai==1.3.8          # For embeddings fallback
groq==0.4.1            # For LLM
transformers==4.36.2   # For local embeddings

# Geographic calculations
geopy==2.4.1

# Data processing
pandas==2.1.4
pydantic==2.5.2

# Logging and monitoring
structlog==23.2.0
